Move submesh functionality into shapes
