+++ modepy/__init__.py
@@ -25,7 +25,8 @@ THE SOFTWARE.
 from modepy.shapes import (
         Shape, Face, Simplex, Hypercube,
 
-        biunit_vertices_for_shape, faces_for_shape
+        biunit_vertices_for_shape, faces_for_shape,
+        submesh_for_shape,
         )
 from modepy.spaces import (
         FunctionSpace, PN, QN, space_for_shape)
@@ -78,6 +79,7 @@ __all__ = [
 
         "Shape", "Face", "Simplex", "Hypercube",
         "biunit_vertices_for_shape", "faces_for_shape",
+        "submesh_for_shape",
 
         "FunctionSpace", "PN", "QN", "space_for_shape",
