+++ modepy/shapes.py
@@ -166,6 +166,10 @@ The order of the vertices in the hypercubes follows binary counting
 in ``tsr`` (i.e. in reverse axis order).
 For example, in 3D, ``A, B, C, D, ...`` is ``000, 001, 010, 011, ...``.
 
+Submeshes
+---------
+.. autofunction:: submesh_for_shape
+
 Redirections to Canonical Names
 -------------------------------
 
@@ -403,5 +407,136 @@ def _(shape: Hypercube):
             for iface, fvi in enumerate(face_vertex_indices)]
 
 # }}}
+
+
+# {{{ submeshes
+
+@singledispatch
+def submesh_for_shape(shape: Shape, node_tuples):
+    """Return a list of tuples of indices into the node list that
+    generate a tesselation of the reference element.
+
+    :arg node_tuples: A list of tuples *(i, j, ...)* of integers
+        indicating node positions inside the unit element. The
+        returned list references indices in this list.
+
+        :func:`modepy.node_tuples_for_space` may be used to generate *node_tuples*.
+
+    .. versionadded:: 2020.3
+    """
+    raise NotImplementedError(type(shape).__name__)
+
+
+@submesh_for_shape.register(Simplex)
+def _(shape: Simplex, node_tuples):
+    from pytools import single_valued, add_tuples
+    dims = single_valued(len(nt) for nt in node_tuples)
+
+    node_dict = {
+            ituple: idx
+            for idx, ituple in enumerate(node_tuples)}
+
+    if dims == 1:
+        result = []
+
+        def try_add_line(d1, d2):
+            try:
+                result.append((
+                    node_dict[add_tuples(current, d1)],
+                    node_dict[add_tuples(current, d2)],
+                    ))
+            except KeyError:
+                pass
+
+        for current in node_tuples:
+            try_add_line((0,), (1,),)
+
+        return result
+    elif dims == 2:
+        # {{{ triangle sub-mesh
+        result = []
+
+        def try_add_tri(d1, d2, d3):
+            try:
+                result.append((
+                    node_dict[add_tuples(current, d1)],
+                    node_dict[add_tuples(current, d2)],
+                    node_dict[add_tuples(current, d3)],
+                    ))
+            except KeyError:
+                pass
+
+        for current in node_tuples:
+            # this is a tesselation of a square into two triangles.
+            # subtriangles that fall outside of the master triangle are
+            # simply not added.
+
+            # positively oriented
+            try_add_tri((0, 0), (1, 0), (0, 1))
+            try_add_tri((1, 0), (1, 1), (0, 1))
+
+        return result
+
+        # }}}
+    elif dims == 3:
+        # {{{ tet sub-mesh
+
+        def try_add_tet(d1, d2, d3, d4):
+            try:
+                result.append((
+                    node_dict[add_tuples(current, d1)],
+                    node_dict[add_tuples(current, d2)],
+                    node_dict[add_tuples(current, d3)],
+                    node_dict[add_tuples(current, d4)],
+                    ))
+            except KeyError:
+                pass
+
+        result = []
+        for current in node_tuples:
+            # this is a tesselation of a cube into six tets.
+            # subtets that fall outside of the master tet are simply not added.
+
+            # positively oriented
+            try_add_tet((0, 0, 0), (1, 0, 0), (0, 1, 0), (0, 0, 1))
+            try_add_tet((1, 0, 1), (1, 0, 0), (0, 0, 1), (0, 1, 0))
+            try_add_tet((1, 0, 1), (0, 1, 1), (0, 1, 0), (0, 0, 1))
+
+            try_add_tet((1, 0, 0), (0, 1, 0), (1, 0, 1), (1, 1, 0))
+            try_add_tet((0, 1, 1), (0, 1, 0), (1, 1, 0), (1, 0, 1))
+            try_add_tet((0, 1, 1), (1, 1, 1), (1, 0, 1), (1, 1, 0))
+
+        return result
+
+        # }}}
+    else:
+        raise NotImplementedError("%d-dimensional sub-meshes" % dims)
+
+
+@submesh_for_shape.register(Hypercube)
+def _(shape: Hypercube, node_tuples):
+    from pytools import single_valued, add_tuples
+    dims = single_valued(len(nt) for nt in node_tuples)
+
+    node_dict = {
+            ituple: idx
+            for idx, ituple in enumerate(node_tuples)}
+
+    from pytools import generate_nonnegative_integer_tuples_below as gnitb
+
+    result = []
+    for current in node_tuples:
+        try:
+            result.append(tuple(
+                    node_dict[add_tuples(current, offset)]
+                    for offset in gnitb(2, dims)))
+
+        except KeyError:
+            pass
+
+    return result
+
+# }}}
+
 
 # vim: foldmethod=marker
